fix: Do not cast ticks function to int [skip ci] (#3150)
fix: Do not cast ticks function to int

+++ common/src/main/java/com/wynntils/functions/MinecraftFunctions.java
@@ -45,10 +45,10 @@ public class MinecraftFunctions {
         }
     }
 
-    public static class TicksFunction extends Function<Integer> {
+    public static class TicksFunction extends Function<Long> {
         @Override
-        public Integer getValue(FunctionArguments arguments) {
-            return (int) McUtils.mc().level.getGameTime();
+        public Long getValue(FunctionArguments arguments) {
+            return McUtils.mc().level.getGameTime();
         }
     }
